papers/Risk_Bounds_Aleatoric_Uncertainty/requirements.txt
cvxpy==1.3.0
matplotlib==3.5.2
numpy==1.22.3
pandas==1.4.2
patsy==0.5.2
pytorch_lightning==1.8.6
PyYAML==6.0
scikit_learn==1.2.1
scipy==1.8.1
torch==1.13.1
